+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.2.0 - 2024-11-26
+
+### Added
+
+- Support the `VersionLabel` response element (as the `version_label` property)
+- Bump boto3 and related libraries
+
 ## 2.1.2 - 2024-06-18
 
 ### Changed